review the backend

+++ application.py
@@ -110,28 +110,36 @@ def books(page=1):
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
-        perpage=200
-        startat=page*perpage-200
+        # Pagination
+        perpage=100
+        startat=page*perpage-perpage
+        # Handle a search session storage
         if request.method == 'POST':
             search = request.form.get("search")
             session['search'] = search
         else:
             search = session.get('search')
-
         # Search SQL
+        books = db.execute("SELECT * FROM books WHERE LOWER(title) LIKE LOWER(:search) OR  LOWER(author) LIKE LOWER(:search) OR  LOWER(isbn) LIKE LOWER(:search) OR  CAST(year AS TEXT) LIKE :search ORDER BY isbn LIMIT :perpage OFFSET :startat",
+                {"search": '%'+search+'%', "perpage": perpage, "startat": startat}).fetchall()
+        # Count SQL
         count = db.execute("SELECT count(*) as nb FROM books WHERE LOWER(title) LIKE LOWER(:search) OR  LOWER(author) LIKE LOWER(:search) OR  LOWER(isbn) LIKE LOWER(:search) OR  CAST(year AS TEXT) LIKE :search",
-            {"search": '%'+search+'%'}).fetchone()
+                {"search": '%'+search+'%'}).fetchone()
         nb_books = count.nb
-
+        # Check if books are returned
         if nb_books == 0:
             nb_pages = 0
             message = 'No book found'
         else:
-            books = db.execute("SELECT * FROM books WHERE LOWER(title) LIKE LOWER(:search) OR  LOWER(author) LIKE LOWER(:search) OR  LOWER(isbn) LIKE LOWER(:search) OR  CAST(year AS TEXT) LIKE :search ORDER BY isbn LIMIT :perpage OFFSET :startat",
-                {"search": '%'+search+'%', "perpage": perpage, "startat": startat}).fetchall()
             nb_pages = math.ceil(nb_books/perpage)
             message = f'Found {nb_books} book(s)'
-        return render_template("books.html", books=books, search=search, message=message, nb_pages=nb_pages, page=page)
+        return render_template("books.html",
+                books=books,
+                search=search,
+                message=message,
+                nb_pages=nb_pages,
+                page=page
+                )
 
 @app.route("/books/author/<string:author>")
 def author(author):
@@ -154,21 +162,22 @@ def book(book_isbn):
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
-
         # Make sure book exists.
         book = db.execute("SELECT * FROM books WHERE isbn = :isbn", {"isbn": book_isbn}).fetchone()
         if book is None:
             return render_template("error.html", message="No such book.")
-
         # Get reviews from googleapis
         gapi = requests.get("https://www.googleapis.com/books/v1/volumes",
                         params={"q": "isbn:"+book_isbn})
         if gapi.status_code != 200:
             raise Exception("ERROR: API request unsuccessful.")
         gdata = gapi.json()
-        description = gdata["items"][0]["volumeInfo"]["description"]
-        language = gdata["items"][0]["volumeInfo"]["language"]
-
+        try:
+            description = gdata["items"][0]["volumeInfo"]["description"]
+            language = gdata["items"][0]["volumeInfo"]["language"]
+        except KeyError:
+            description=""
+            language=""
         # Get reviews from goodread API
         res = requests.get("https://www.goodreads.com/book/review_counts.json",
                         params={"key": "GtG9odDoZNEWxekOhsmMA", "isbns": book_isbn})
@@ -177,15 +186,13 @@ def book(book_isbn):
         data = res.json()
         ratings_count = data["books"][0]["ratings_count"]
         average_rating = data["books"][0]["average_rating"]
-
         # Get some usefull variables
         user_name = session.get('user_name') or None
         user_id = session.get('user_id') or None
         review_by_user = db.execute("SELECT * FROM reviews WHERE isbn = :isbn AND user_id = :user_id", {"isbn": book_isbn, "user_id": user_id}).fetchone()
-
         # Insert a new review
         if request.method == 'POST':
-            # Get form information.
+            # Get form information
             rating = request.form.get("rating") or None
             content = request.form.get("content") or None
             if review_by_user:
@@ -197,20 +204,19 @@ def book(book_isbn):
                 review_by_user = True
             else:
                 flash("Please correct data")
-
         # Get reviews for the book
         reviews = db.execute("SELECT reviews.id, to_char(reviews.date, 'DD/MM/YY') as date, reviews.rating, reviews.content, users.name FROM reviews JOIN users ON users.id = reviews.user_id WHERE isbn = :isbn", {"isbn": book_isbn}).fetchall()
-
+        # Render the book page
         return render_template("book.html",
-                        book=book,
-                        reviews=reviews,
-                        ratings_count=ratings_count,
-                        average_rating=average_rating,
-                        review_by_user=review_by_user,
-                        user_name=user_name,
-                        description=description,
-                        language=language
-                        )
+                book=book,
+                reviews=reviews,
+                ratings_count=ratings_count,
+                average_rating=average_rating,
+                review_by_user=review_by_user,
+                user_name=user_name,
+                description=description,
+                language=language
+                )
 
 @app.route("/delete_review", methods=["POST"])
 def delete_review():
@@ -219,48 +225,52 @@ def delete_review():
     else:
         book_isbn = request.form.get("book_isbn") or None
         user_id = session.get('user_id')
-
-        db.execute("DELETE FROM reviews WHERE user_id = :user_id AND isbn = :isbn",
-            {"user_id": user_id, "isbn": book_isbn})
-        db.commit()
-        return redirect(url_for('book', book_isbn=book_isbn))
+        if all(v is not None for v in [book_isbn, user_id]):
+            db.execute("DELETE FROM reviews WHERE user_id = :user_id AND isbn = :isbn",
+                {"user_id": user_id, "isbn": book_isbn})
+            db.commit()
+            return redirect(url_for('book', book_isbn=book_isbn))
+        else:
+            return render_template("error.html", message="Please check your review.")
 
 @app.route("/review/<int:review_id>", methods=["GET", "POST"])
 def update_review(review_id):
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
-        
         # Make sure review exists.
         review = db.execute("SELECT * FROM reviews WHERE id = :id", {"id": review_id}).fetchone()
         if review is None:
             return render_template("error.html", message="No such review.")
-        
+        # Get the isbn and user
         book_isbn = review.isbn
         user_id = session.get('user_id')
-
+        # Check if the user can update the review
         if user_id != review.user_id:
             return render_template("error.html", message="Not your review.")
+        # Update the review
         elif request.method == 'POST':
-            book_isbn = request.form.get("isbn") or None
             rating = request.form.get("rating") or None
             content = request.form.get("content") or None
-
-            db.execute("UPDATE reviews SET date = current_timestamp, rating = :rating, content = :content WHERE user_id = :user_id AND isbn = :isbn",
-                {"rating": rating, "content": content, "user_id": user_id, "isbn": book_isbn})
-            db.commit()
-            return redirect(url_for('book', book_isbn=book_isbn))
+            if all(v is not None for v in [rating, content]):
+                db.execute("UPDATE reviews SET date = current_timestamp, rating = :rating, content = :content WHERE user_id = :user_id AND isbn = :isbn",
+                    {"rating": rating, "content": content, "user_id": user_id, "isbn": book_isbn})
+                db.commit()
+                return redirect(url_for('book', book_isbn=book_isbn))
+            else:
+                flash("Please add rating and content.")
         else:
             review = db.execute("SELECT reviews.id, reviews.date, reviews.rating, reviews.isbn, reviews.content, books.title, users.name FROM reviews JOIN books ON books.isbn = reviews.isbn JOIN users ON users.id = reviews.user_id WHERE reviews.id = :review_id AND reviews.user_id = :user_id", {"review_id": review_id, "user_id": user_id}).fetchone()
-            return render_template('update_review.html', review=review)
+        return render_template('update_review.html', review=review)
 
+# List reviews (For admin only)
 @app.route("/reviews")
 def reviews():
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
         reviews = db.execute("SELECT reviews.id, to_char(reviews.date, 'DD/MM/YY') as date, reviews.rating, reviews.content, reviews.user_id, users.name, books.title,  books.isbn FROM reviews JOIN users ON users.id = reviews.user_id JOIN books ON books.isbn = reviews.isbn").fetchall()
-        # Check if user has riviews
+        # Check if reviews exist
         if reviews:
             nb_reviews = len(reviews) or None
             message = f'Found {nb_reviews} reviews'
@@ -268,13 +278,14 @@ def reviews():
         else:
             return render_template("error.html", message="No review for this user.")
 
+# List reviews by user (For admin only)
 @app.route("/reviews/<int:user_id>")
 def reviews_by_user(user_id):
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
         reviews = db.execute("SELECT reviews.id, to_char(reviews.date, 'DD/MM/YY') as date, reviews.rating, reviews.content, reviews.user_id, users.name, books.title,  books.isbn FROM reviews JOIN users ON users.id = reviews.user_id JOIN books ON books.isbn = reviews.isbn WHERE reviews.user_id = :user_id", {"user_id": user_id}).fetchall()
-         # Check if user has riviews
+         # Check if reviews exist
         if reviews:
             username = db.execute("SELECT users.name FROM users JOIN reviews ON users.id = reviews.user_id WHERE reviews.user_id = :user_id", {"user_id": user_id}).fetchone()
             nb_reviews = len(reviews) or None
@@ -283,6 +294,7 @@ def reviews_by_user(user_id):
         else:
             return render_template("error.html", message="No review for this user.")
 
+# List users (For admin only)
 @app.route("/users", methods=['GET', 'POST'])
 def users():
     if not session.get('logged_in'):
@@ -290,61 +302,63 @@ def users():
     elif not session.get('user_isadmin'):
         return render_template('error.html', message='For admin users only')
     else:
+        # Create a new user
         if request.method == 'POST':
             # Get form information.
             username = str(request.form.get("username")) or None
             email = str(request.form.get("email")) or None
             password = str(request.form.get("password")) or None
-
+            # Validation
             if all(v is not None for v in [username, email, password]):
                 db.execute("INSERT INTO users (name, email, password) VALUES (:username, :email, :password)",
                 {"username": username, "email": email, "password": password})
                 db.commit()
-                users = db.execute("SELECT * FROM users").fetchall()
-                return render_template("users.html", users=users)
             else:
-                message = "No empty field"
-                users = db.execute("SELECT * FROM users").fetchall()
-                return render_template("users.html", users=users, message=message)
-        else:
-            users = db.execute("SELECT * FROM users").fetchall()
-            nb_users = len(users)
-            message = f'Found {nb_users} users'
-            return render_template("users.html", users=users, message=message)
+                flash("No empty field")
+        # List the users
+        users = db.execute("SELECT * FROM users").fetchall()
+        nb_users = len(users)
+        message = f'Found {nb_users} users'
+        return render_template("users.html", users=users, message=message)
 
 @app.route("/user/<int:user_id>", methods=['GET', 'POST'])
 def profile(user_id):
     if not session.get('logged_in'):
         return render_template('login.html')
     else:
+        # Only display user's profile
         current_user_id = session.get("user_id") or None
         if user_id != current_user_id:
             return render_template("error.html", message="Not you.")
-
+        # Get the users reviews
         reviews = db.execute("SELECT reviews.id, to_char(reviews.date, 'DD/MM/YY') as date, reviews.rating, reviews.content, reviews.user_id, users.name, books.title,  books.isbn FROM reviews JOIN users ON users.id = reviews.user_id JOIN books ON books.isbn = reviews.isbn WHERE reviews.user_id = :user_id",
                 {"user_id": user_id}).fetchall()
+        # User can update infos
         if request.method == 'POST':
             username = request.form.get("username") or None
             email = request.form.get("email") or None
             password = request.form.get("password") or None
-
-            db.execute("UPDATE users SET name = :username, email = :email, password = :password WHERE id = :user_id",
-                {"username": username, "email": email, "password": password, "user_id": user_id})
-            db.commit()
-            user = db.execute("SELECT * FROM users WHERE id = :user_id", {"user_id": user_id}).fetchone()
-            flash("You've updated your infos")
-            return render_template("profile.html", user=user, reviews=reviews)
-        else:
-            user = db.execute("SELECT * FROM users WHERE id = :user_id", {"user_id": user_id}).fetchone()
-            return render_template("profile.html", user=user, reviews=reviews)
+            # Validation
+            if all(v is not None for v in [username, email, password]):
+                db.execute("UPDATE users SET name = :username, email = :email, password = :password WHERE id = :user_id",
+                    {"username": username, "email": email, "password": password, "user_id": user_id})
+                db.commit()
+                flash("You've updated your infos")
+            else:
+                flash("No empty field")
+        # Get user info
+        user = db.execute("SELECT * FROM users WHERE id = :user_id", {"user_id": user_id}).fetchone()
+        return render_template("profile.html", user=user, reviews=reviews)
 
 @app.route("/api/<string:book_isbn>")
 def api(book_isbn):
+    # Retrieve infos
     book = db.execute("SELECT * FROM books WHERE isbn = :book_isbn",
             {"book_isbn": book_isbn}).fetchone()
+    # Check if book exists in the db
     if book is None:
-        return jsonify({"error": "Invalid ISBN"}), 422
-
+        # return jsonify({"error": "Invalid ISBN"}), 422
+        return render_template("error.html", message="No such book.")
     # Get reviews from goodread API
     res = requests.get("https://www.goodreads.com/book/review_counts.json",
                     params={"key": "GtG9odDoZNEWxekOhsmMA", "isbns": book_isbn})
@@ -353,7 +367,7 @@ def api(book_isbn):
     data = res.json()
     ratings_count = data["books"][0]["ratings_count"]
     average_rating = data["books"][0]["average_rating"]
-
+    # Make the return
     return jsonify({
             "title": book.title,
             "author": book.author,
